upcl: 8086 0xF7 /digit group (test/not/neg/mul/imul/div/idiv) + fix interp signed division
Add the second grouped opcode whose ModR/M `reg` field is an opcode
extension: 0xF7 /digit (word):

    /0 test (imm16, a tail immediate)   /2 not   /3 neg
    /4 mul   /5 imul   /6 div   /7 idiv

MUL/DIV use the implicit AX (and DX for the upper half). The 16x16 product
is 32 bits, so the operands are widened ([ #i32 .. ] zero-extends, %S
sign-extends) before the multiply and the result is sliced back into DX:AX;
the 32-bit dividend DX:AX is built with the bit-combine ( dx : ax ), divided,
and the quotient/remainder truncated into AX/DX. No new language constructs
-- this exercises the existing cast, %S signedness and bit-combine forms.

Fix a real interpreter-backend bug this surfaced: BinSDiv / BinSRem divided
the SIGN-EXTENDED operands with a UINT64 `/`, which is unsigned -- it divided
the full 64-bit magnitudes (2^64-n) / m and kept the low bits, so a signed
-20 / 3 produced 0x554E instead of 0xFFFA (-6). Reinterpret the sign-extended
patterns as INT64 before dividing. (The neighbouring BinAShr is unaffected:
a logical shift of the sign-extended value, then masking, equals an arithmetic
shift -- division has no such self-correction.) The LLVM backend lowered
BinSDiv correctly all along, so only the default interpreter was wrong, and
signed division had never been exercised by a passing test until now.

Test upcl.x86-f7group decodes the group and runs the signed IDIV (-20/3 ->
ax=0xFFFA, dx=0xFFFE) as the BinSDiv regression guard, plus MUL (-> DX:AX).
232/232 tests pass.

+++ LibCPU/backends/interp/Interp.cpp
@@ -268,9 +268,12 @@ private:
         case BinSub:  R = A - B; break;
         case BinMul:  R = A * B; break;
         case BinUDiv: R = B ? A / B : 0; break;
-        case BinSDiv: R = B ? (UINT64)(SignExtend (A, Bits) / SignExtend (B, Bits)) : 0; break;
+        // Signed divide/remainder: the operands are sign-extended bit patterns, but the operator
+        // itself must be signed too -- a UINT64 `/` would divide their full 64-bit magnitudes
+        // unsigned (2^64-n / m), so reinterpret as INT64 before dividing.
+        case BinSDiv: R = B ? (UINT64)((INT64) SignExtend (A, Bits) / (INT64) SignExtend (B, Bits)) : 0; break;
         case BinURem: R = B ? A % B : 0; break;
-        case BinSRem: R = B ? (UINT64)(SignExtend (A, Bits) % SignExtend (B, Bits)) : 0; break;
+        case BinSRem: R = B ? (UINT64)((INT64) SignExtend (A, Bits) % (INT64) SignExtend (B, Bits)) : 0; break;
         case BinAnd:  R = A & B; break;
         case BinOr:   R = A | B; break;
         case BinXor:  R = A ^ B; break;
